requirements.txt
python-telegram-bot
peewee
tweepy